EPMRPP-33456: Code review fixes - 1, split launchDurationTooltip to separate folder
