+++ app/src/components/widgets/charts/launchesDurationChart/launchDurationTooltip/index.js
@@ -1,0 +1,1 @@
+export { LaunchDurationTooltip } from './launchDurationTooltip';
