+++ app/src/components/widgets/charts/launchesDurationChart/launchDurationTooltip/launchDurationTooltip.scss
@@ -1,0 +1,12 @@
+.launch-name {
+  width: 100%;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  font-family: $FONT-SEMIBOLD, sans-serif;
+}
+
+.launch-duration {
+  width: 100%;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
