+++ app/src/components/widgets/charts/launchesDurationChart/launchesDurationChart.jsx
@@ -7,7 +7,7 @@ import { COLOR_CHART_DURATION, COLOR_FAILED } from 'common/constants/colors';
 import { COMMON_LOCALE_KEYS } from 'common/constants/localization';
 
 import { C3Chart } from '../common/c3chart';
-import { LaunchDurationTooltip } from './tooltip';
+import { LaunchDurationTooltip } from './launchDurationTooltip';
 import { isValueInterrupted, transformCategoryLabel, getLaunchAxisTicks } from './chartUtils';
 import { prepareChartData } from './prepareChartData';
 import { DURATION } from './constants';
